+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
 - **Tables Overview**:
   - Added optional column "Tutorials" also to the regular table overview. This way, users can easier see if there will be a tutorial available in the pause menu.
+  - Improved ROM column: The tooltip of a ROM displays additional alias information now. In addition to that, you can use the table search field to filter games by their ROM name or alias.
 - **Pause Menu**
   - Added pause menu support for stand-alone mode of VPX.